Convert the synthsizer to an iterative mixer
In order to allow custom sampled waveforms, the mixer need to be
iterative.  For every entry in the active notes LUT, the sample is added
to the total mix.  This happens right before a new sample needs to be
generated, in 128 clock cycles.  A square wave is still used as output,
but this is now very easy to convert to a sample-based synthesizer.
Samples will need to be represented in signed PCM, using 2-complement
binary representation.